Consolidate data category calls
